requirements.txt
absl-py==0.7.0
ftfy==5.8
gdown==3.13.1
gym==0.17.3
hydra-core==1.0.5
imageio==2.9.0
imageio-ffmpeg==0.4.5
kornia==0.4.1
matplotlib==3.3.4
meshcat==0.0.18
numpy==1.19.5
opencv-python==4.5.5.64
pybullet==3.0.4
pytorch_lightning==1.0.3
regex==2021.4.4
scikit-image==0.17.2
scipy==1.4.1
tqdm==4.62.3
torch==1.9.1+cu111
torchvision==0.10.1+cu111
transformers==4.3.2
transforms3d==0.3.1
wandb==0.10.15
